Revert "Move package to org.http4s.scalaxml1 (#1)"
This reverts commit ba79975a0d92e37db3699eb967c5fc374672299f.
