+++ scala-xml/src/main/scala/scalaxml/ElemInstances.scala
@@ -15,7 +15,7 @@
  */
 
 package org.http4s
-package scalaxml1
+package scalaxml
 
 import cats.effect.Concurrent
 import org.http4s.Charset.`UTF-8`
